requirements.txt
asgiref==3.3.4
Cerberus==1.3.3
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.7.0
django-heroku==0.3.1
django-extensions==3.0.6
djangorestframework==3.12.4
gunicorn==19.9.0
psycopg2==2.8.6
PyJWT==1.7.1
python-dotenv==0.17.0
python-dateutil==2.8.1
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
